[XNNPACK][Partitioner] Migrate completely to new config based partitioner
New Config Based Partitioner maintains parity with old partitioners. I believe we can now replace all the old partitioners with the new partitioners

I also added pre configured partitioners like the ones most users use today, so that this would not cause any problems with people now. Despite having "multiple" partitioners It is the same partitioenr with different specified configs. The new xnnpack_partitioner.py file contains a lot of changes since it is essentially getting deleted and replaced with xnnpack_partitioner.py. But if you just look at the new file itself it is pretty straight forward.

https://www.internalfb.com/code/fbsource/[D61250577-V2]/fbcode/executorch/backends/xnnpack/partition/xnnpack_partitioner.py

Differential Revision: [D61250577](https://our.internmc.facebook.com/intern/diff/D61250577/)

[ghstack-poisoned]
